A-010-P2: promote settings to full-page with background reconnect toggle
